% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_TwinBKP.R
\name{.twin_2D_plot_fun}
\alias{.twin_2D_plot_fun}
\title{Internal 2D plotting helper for TwinBKP}
\usage{
.twin_2D_plot_fun(
  var,
  title,
  data,
  X = NULL,
  y = NULL,
  X_global = NULL,
  dims = NULL,
  ...
)
}
\description{
Internal 2D plotting helper for TwinBKP
}
\keyword{internal}
