% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_prior_dkp.R
\name{get_prior_dkp}
\alias{get_prior_dkp}
\title{Construct Prior Parameters for Dirichlet Kernel Process (DKP)}
\usage{
get_prior_dkp(
  prior = c("noninformative", "fixed", "adaptive"),
  r0,
  p0,
  Y,
  K = NULL
)
}
\arguments{
\item{prior}{Type of prior: one of \code{"noninformative"}, \code{"fixed"}, or \code{"adaptive"}.}

\item{r0}{Global precision parameter (used in \code{"fixed"} and \code{"adaptive"}).}

\item{p0}{Global prior mean used when prior = "fixed".}

\item{Y}{A numeric vector of observed successes.}

\item{K}{A precomputed kernel matrix (n x n), typically from \code{kernel_matrix()}.}
}
\value{
A list with a numeric vectors: \code{alpha0} (of length n).
}
\description{
Constructs prior Beta distribution parameters (alpha0 and beta0) at each location
based on user-specified prior type: noninformative, fixed, or data-adaptive.
}
\examples{
get_prior_dkp(prior = "noninformative", Y = rep(1,3))
}
