% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_fun_dkp.R
\name{lossFun}
\alias{lossFun}
\alias{loss_fun_dkp}
\title{Loss Function for DKP Model Fitting}
\usage{
loss_fun_dkp(
  gamma,
  Xnorm,
  Y,
  prior = c("noninformative", "fixed", "adaptive"),
  r0 = 2,
  p0 = NULL,
  loss = c("brier", "log_loss"),
  kernel = c("gaussian", "matern52", "matern32")
)
}
\arguments{
\item{gamma}{A numeric vector of kernel hyperparameters (on the log scale).}

\item{Xnorm}{A normalized input matrix (each column scaled \code{[0,1]}).}

\item{Y}{A matrix of observed successes of size \eqn{n \times q}.}

\item{prior}{Type of prior: "noninformative", "fixed", or "adaptive".}

\item{r0}{Global prior strength (used in "fixed"/"adaptive").}

\item{p0}{Prior mean for fixed prior.}

\item{loss}{Type of loss: "brier" or "log_loss".}

\item{kernel}{Type of kernel: "gaussian", "matern52", or "matern32".}
}
\value{
A scalar numeric value representing the loss.
}
\description{
Computes the loss used for fitting the Beta Kernel Process (DKP) model.
Supports Brier score (mean squared error) or log-loss (cross-entropy),
under different prior specifications.
}
\examples{
loss_fun_dkp(gamma = rep(1, 2),
        Xnorm = matrix(runif(20), ncol=2),
        Y = matrix(runif(20), ncol=2),
        prior = "noninformative",
        loss = "brier",
        kernel = "gaussian")
}
\author{
Jiangyan Zhao, Kunhai Qing, Jin Xu
}
