% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twining.R
\name{get_twin_indices}
\alias{get_twin_indices}
\title{Internal Twin index selector}
\usage{
get_twin_indices(Xnorm, g, v = 2L * g, runs = 10L, seed = 123L)
}
\description{
Internal Twin index selector
}
\keyword{internal}
