% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loss_fun.R
\name{loss_fun}
\alias{loss_fun}
\title{Loss Function for BKP Model Fitting}
\usage{
loss_fun(
  gamma,
  Xnorm,
  y,
  m,
  prior = c("noninformative", "fixed", "adaptive"),
  r0 = 2,
  p0 = 0.5,
  loss = c("brier", "log_loss"),
  kernel = c("gaussian", "matern52", "matern32")
)
}
\arguments{
\item{gamma}{A numeric vector of kernel hyperparameters (on the log scale).}

\item{Xnorm}{A normalized input matrix (each column scaled \code{[0,1]}).}

\item{y}{A numeric vector of binary counts (successes).}

\item{m}{A numeric vector of total trials (denominators).}

\item{prior}{Type of prior: "noninformative", "fixed", or "adaptive".}

\item{r0}{Global prior strength (used in "fixed"/"adaptive").}

\item{p0}{Prior mean for fixed prior (in (0,1)).}

\item{loss}{Type of loss: "brier" or "log_loss".}

\item{kernel}{Type of kernel: "gaussian", "matern52", or "matern32".}
}
\value{
A scalar numeric value representing the loss.
}
\description{
Computes the loss used for fitting the Beta Kernel Process (BKP)
model. Supports Brier score (mean squared error) or log-loss
(cross-entropy), under different prior specifications.
}
\examples{
loss_fun(gamma = rep(0, 2),
        Xnorm = matrix(runif(20), ncol=2),
        y = rbinom(10, 10, 0.5),
        m = rep(10, 10))

}
\author{
Jiangyan Zhao, Kunhai Qing, Jin Xu
}
