% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictBKP.R
\name{predict}
\alias{predict}
\alias{predict.BKP}
\title{Predict method for Beta Kernel Process (BKP) models}
\usage{
\method{predict}{BKP}(object, Xnew, CI_size = 0.05)
}
\arguments{
\item{object}{A fitted BKP model object returned by \code{\link{fit.BKP}}.}

\item{Xnew}{A matrix (or vector) of new input points at which to predict.}

\item{CI_size}{Confidence interval level (default = 0.05 for 95\% CI).}
}
\value{
A data frame with columns:
\describe{
\item{X}{Original new input locations}
\item{mean}{Posterior mean of success probability}
\item{variance}{Posterior variance}
\item{lower}{Lower bound of CI for success probability}
\item{upper}{Upper bound of CI for success probability}
}
}
\description{
Generate predictions from a fitted BKP model at new input locations.
}
\examples{
### 1D
set.seed(123)
n <- 30
Xbounds <- matrix(c(-2,2), nrow=1)
x <- seq(-2, 2, length = n)
true_pi <- (1 + exp(-x^2) * cos(10 * (1 - exp(-x)) / (1 + exp(-x)))) / 2
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)
df <- data.frame(x = x, y = y, m = m)
xx = matrix(seq(-2, 2, length = 100), ncol=1) #new data points
model <- fit.BKP(df, Xbounds=Xbounds)
head(predict(model,xx))

### 2D
set.seed(123)
n <- 100
f <- function(X) {
  if(is.null(nrow(X))) X <- matrix(X, nrow=1)
  m <- 8.6928
  s <- 2.4269
  x1 <- 4*X[,1]- 2
  x2 <- 4*X[,2]- 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19- 14*x1 + 3*x1^2- 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1- 3*x2)^2 *
    (18- 32*x1 + 12*x1^2 + 48*x2- 36*x1*x2 + 27*x2^2)
  f <- log(a*b)
  f <- (f- m)/s
  return(f) }
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
x <- tgp::lhs(n = n, rect = Xbounds)
true_pi <- pnorm(f(x))
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)
df <- data.frame(x = x, y = y, m = m)
xx1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 100)
xx2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 100)
xx <- expand.grid(xx1 = xx1, xx2 = xx2)
model <- fit.BKP(df)
head(predict(model,xx))

}
\author{
Jiangyan Zhao, Kunhai Qing, Jin Xu
}
\keyword{BKP}
