% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_BKP.R
\name{print}
\alias{print}
\alias{print.BKP}
\title{Print Summary of a BKP Model}
\usage{
\method{print}{BKP}(x, ...)
}
\arguments{
\item{x}{An object of class \code{"BKP"}, typically returned by \code{\link{fit.BKP}}.}

\item{...}{Additional arguments passed to the generic \code{print} method (not used here).}
}
\description{
Provides a concise summary of a fitted Beta Kernel Process (BKP) model object.
The printed output includes key model characteristics such as the number of observations,
the dimensionality of the input, the kernel and loss type used, the optimized kernel
parameters (\code{bestTheta}), and the minimum loss (e.g., Brier score) achieved.
}
\details{
When a \code{BKP} object is printed (e.g., by calling \code{print(model)} or typing the object name
in the console), this method presents a readable summary of the fitted model's essential properties.
}
\examples{
\dontrun{
### 1D
set.seed(123)
n <- 100
Xbounds <- matrix(c(-2,2), nrow=1)
x <- seq(-2, 2, length = n)
true_pi <- (1 + exp(-x^2) * cos(10 * (1 - exp(-x)) / (1 + exp(-x)))) / 2
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)
df <- data.frame(x = x, y = y, m = m)
xx = matrix(seq(-2, 2, length = 100), ncol=1) #new data points
model <- fit.BKP(df, Xbounds=Xbounds)
print(model)

### 2D
set.seed(123)
n <- 100
f <- function(X) {
  if(is.null(nrow(X))) X <- matrix(X, nrow=1)
  m <- 8.6928
  s <- 2.4269
  x1 <- 4*X[,1]- 2
  x2 <- 4*X[,2]- 2
  a <- 1 + (x1 + x2 + 1)^2 *
    (19- 14*x1 + 3*x1^2- 14*x2 + 6*x1*x2 + 3*x2^2)
  b <- 30 + (2*x1- 3*x2)^2 *
    (18- 32*x1 + 12*x1^2 + 48*x2- 36*x1*x2 + 27*x2^2)
  f <- log(a*b)
  f <- (f- m)/s
  return(f) }
Xbounds <- matrix(c(0, 0, 1, 1), nrow = 2)
library(tgp)
x <- lhs(n = n, rect = Xbounds)
true_pi <- pnorm(f(x))
m <- sample(100, n, replace = TRUE)
y <- rbinom(n, size = m, prob = true_pi)
df <- data.frame(x = x, y = y, m = m)
xx1 <- seq(Xbounds[1,1], Xbounds[1,2], length.out = 100)
xx2 <- seq(Xbounds[2,1], Xbounds[2,2], length.out = 100)
xx <- expand.grid(xx1 = xx1, xx2 = xx2)
model <- fit.BKP(df)
print(model)
}

}
\seealso{
\code{\link{fit.BKP}} for model fitting.
\code{\link{summary.BKP}} for a more detailed report, if implemented.
}
\author{
Jiangyan Zhao, Kunhai Qing, Jin Xu
}
\keyword{BKP}
